Change _tsarray_priv.capacity and _tsarray_priv.len to unsigned long.
This is a breaking change, part of a long series of small commits that
will not build. We're doing a massive change in small parts.

[skip ci]

+++ src/tsarray.c
@@ -62,8 +62,8 @@
 struct _tsarray_priv {
     struct _tsarray_pub pub;    /* MUST be the first member (read above) */
     size_t obj_size;
-    size_t capacity;
-    size_t len;
+    unsigned long capacity;     /* keep <= LONG_MAX, as indices are signed */
+    unsigned long len;          /* likewise */
 };
